Refactored routes into separate folder
