+++ api/routes/questionRoutes.js
@@ -1,0 +1,25 @@
+const express = require('express')
+const knex = require('knex')(require('../knexfile'))
+
+const router = express.Router()
+
+router.get('/', async (req, res) => {
+  try {
+    const questions = await knex.select().table('questions')
+    res.json(questions)
+  } catch (err) {
+    res.status(500)
+  }
+})
+  
+router.post('/', async (req, res) => {
+  const { text } = req.body
+  try {
+    const question = await knex('questions').insert({ text }, '*')
+    res.json(question)
+  } catch (err) {
+    res.status(text ? 500 : 400)
+  }
+})
+
+module.exports = router
